+++ x/distribution/keeper/validator.go
@@ -13,34 +13,64 @@ import (
 
 // initialize rewards for a new validator
 func (k Keeper) initializeValidator(ctx context.Context, val stakingtypes.ValidatorI) error {
+	valBz, err := k.stakingKeeper.ValidatorAddressCodec().StringToBytes(val.GetOperator())
+	if err != nil {
+		return err
+	}
 	// set initial historical rewards (period 0) with reference count of 1
+<<<<<<< HEAD
 	err := k.SetValidatorHistoricalRewards(ctx, val.GetOperator(), 0, types.NewValidatorHistoricalRewards(sdk.DecCoins{}, 1))
+=======
+	err = k.ValidatorHistoricalRewards.Set(ctx, collections.Join(sdk.ValAddress(valBz), uint64(0)), types.NewValidatorHistoricalRewards(sdk.DecCoins{}, 1))
+>>>>>>> e60c583d2 (refactor: migrate away from using valBech32 globals (2/2) (#17157))
 	if err != nil {
 		return err
 	}
 
 	// set current rewards (starting at period 1)
+<<<<<<< HEAD
 	err = k.SetValidatorCurrentRewards(ctx, val.GetOperator(), types.NewValidatorCurrentRewards(sdk.DecCoins{}, 1))
+=======
+	err = k.ValidatorCurrentRewards.Set(ctx, valBz, types.NewValidatorCurrentRewards(sdk.DecCoins{}, 1))
+>>>>>>> e60c583d2 (refactor: migrate away from using valBech32 globals (2/2) (#17157))
 	if err != nil {
 		return err
 	}
 
 	// set accumulated commission
+<<<<<<< HEAD
 	err = k.SetValidatorAccumulatedCommission(ctx, val.GetOperator(), types.InitialValidatorAccumulatedCommission())
+=======
+	err = k.ValidatorsAccumulatedCommission.Set(ctx, valBz, types.InitialValidatorAccumulatedCommission())
+>>>>>>> e60c583d2 (refactor: migrate away from using valBech32 globals (2/2) (#17157))
 	if err != nil {
 		return err
 	}
 
 	// set outstanding rewards
+<<<<<<< HEAD
 	err = k.SetValidatorOutstandingRewards(ctx, val.GetOperator(), types.ValidatorOutstandingRewards{Rewards: sdk.DecCoins{}})
+=======
+	err = k.ValidatorOutstandingRewards.Set(ctx, valBz, types.ValidatorOutstandingRewards{Rewards: sdk.DecCoins{}})
+>>>>>>> e60c583d2 (refactor: migrate away from using valBech32 globals (2/2) (#17157))
 	return err
 }
 
 // increment validator period, returning the period just ended
 func (k Keeper) IncrementValidatorPeriod(ctx context.Context, val stakingtypes.ValidatorI) (uint64, error) {
+	valBz, err := k.stakingKeeper.ValidatorAddressCodec().StringToBytes(val.GetOperator())
+	if err != nil {
+		return 0, err
+	}
+
 	// fetch current rewards
+<<<<<<< HEAD
 	rewards, err := k.GetValidatorCurrentRewards(ctx, val.GetOperator())
 	if err != nil {
+=======
+	rewards, err := k.ValidatorCurrentRewards.Get(ctx, valBz)
+	if err != nil && !errors.Is(err, collections.ErrNotFound) {
+>>>>>>> e60c583d2 (refactor: migrate away from using valBech32 globals (2/2) (#17157))
 		return 0, err
 	}
 
@@ -55,8 +85,13 @@ func (k Keeper) IncrementValidatorPeriod(ctx context.Context, val stakingtypes.V
 			return 0, err
 		}
 
+<<<<<<< HEAD
 		outstanding, err := k.GetValidatorOutstandingRewards(ctx, val.GetOperator())
 		if err != nil {
+=======
+		outstanding, err := k.ValidatorOutstandingRewards.Get(ctx, valBz)
+		if err != nil && !errors.Is(err, collections.ErrNotFound) {
+>>>>>>> e60c583d2 (refactor: migrate away from using valBech32 globals (2/2) (#17157))
 			return 0, err
 		}
 
@@ -67,7 +102,11 @@ func (k Keeper) IncrementValidatorPeriod(ctx context.Context, val stakingtypes.V
 			return 0, err
 		}
 
+<<<<<<< HEAD
 		err = k.SetValidatorOutstandingRewards(ctx, val.GetOperator(), outstanding)
+=======
+		err = k.ValidatorOutstandingRewards.Set(ctx, valBz, outstanding)
+>>>>>>> e60c583d2 (refactor: migrate away from using valBech32 globals (2/2) (#17157))
 		if err != nil {
 			return 0, err
 		}
@@ -79,7 +118,11 @@ func (k Keeper) IncrementValidatorPeriod(ctx context.Context, val stakingtypes.V
 	}
 
 	// fetch historical rewards for last period
+<<<<<<< HEAD
 	historical, err := k.GetValidatorHistoricalRewards(ctx, val.GetOperator(), rewards.Period-1)
+=======
+	historical, err := k.ValidatorHistoricalRewards.Get(ctx, collections.Join(sdk.ValAddress(valBz), rewards.Period-1))
+>>>>>>> e60c583d2 (refactor: migrate away from using valBech32 globals (2/2) (#17157))
 	if err != nil {
 		return 0, err
 	}
@@ -87,19 +130,27 @@ func (k Keeper) IncrementValidatorPeriod(ctx context.Context, val stakingtypes.V
 	cumRewardRatio := historical.CumulativeRewardRatio
 
 	// decrement reference count
-	err = k.decrementReferenceCount(ctx, val.GetOperator(), rewards.Period-1)
+	err = k.decrementReferenceCount(ctx, valBz, rewards.Period-1)
 	if err != nil {
 		return 0, err
 	}
 
 	// set new historical rewards with reference count of 1
+<<<<<<< HEAD
 	err = k.SetValidatorHistoricalRewards(ctx, val.GetOperator(), rewards.Period, types.NewValidatorHistoricalRewards(cumRewardRatio.Add(current...), 1))
+=======
+	err = k.ValidatorHistoricalRewards.Set(ctx, collections.Join(sdk.ValAddress(valBz), rewards.Period), types.NewValidatorHistoricalRewards(cumRewardRatio.Add(current...), 1))
+>>>>>>> e60c583d2 (refactor: migrate away from using valBech32 globals (2/2) (#17157))
 	if err != nil {
 		return 0, err
 	}
 
 	// set current rewards, incrementing period by 1
+<<<<<<< HEAD
 	err = k.SetValidatorCurrentRewards(ctx, val.GetOperator(), types.NewValidatorCurrentRewards(sdk.DecCoins{}, rewards.Period+1))
+=======
+	err = k.ValidatorCurrentRewards.Set(ctx, valBz, types.NewValidatorCurrentRewards(sdk.DecCoins{}, rewards.Period+1))
+>>>>>>> e60c583d2 (refactor: migrate away from using valBech32 globals (2/2) (#17157))
 	if err != nil {
 		return 0, err
 	}